# openapi.json
{"openapi":"3.1.0","info":{"title":"MATA Backend","version":"0.1.0"},"paths":{"/api/v1/admin/upload/rdb":{"post":{"tags":["admin"],"summary":"Upload Rdb","operationId":"upload_rdb_api_v1_admin_upload_rdb_post","parameters":[{"name":"X-User-Role","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Role"}},{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Id"}},{"name":"X-User-Programme","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Programme"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_rdb_api_v1_admin_upload_rdb_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Upload Rdb Api V1 Admin Upload Rdb Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/admin/upload/ttf":{"post":{"tags":["admin"],"summary":"Upload Ttf","operationId":"upload_ttf_api_v1_admin_upload_ttf_post","parameters":[{"name":"X-User-Role","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Role"}},{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Id"}},{"name":"X-User-Programme","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Programme"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_ttf_api_v1_admin_upload_ttf_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Upload Ttf Api V1 Admin Upload Ttf Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/admin/upload/form-f1":{"post":{"tags":["admin"],"summary":"Upload Formf1","operationId":"upload_formf1_api_v1_admin_upload_form_f1_post","parameters":[{"name":"X-User-Role","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Role"}},{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Id"}},{"name":"X-User-Programme","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Programme"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_formf1_api_v1_admin_upload_form_f1_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Upload Formf1 Api V1 Admin Upload Form F1 Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/admin/upload/public-holidays":{"post":{"tags":["admin"],"summary":"Upload Public Holidays","operationId":"upload_public_holidays_api_v1_admin_upload_public_holidays_post","parameters":[{"name":"X-User-Role","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Role"}},{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Id"}},{"name":"X-User-Programme","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Programme"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_public_holidays_api_v1_admin_upload_public_holidays_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Upload Public Holidays Api V1 Admin Upload Public Holidays Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"tags":["health"],"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}}},"components":{"schemas":{"Body_upload_formf1_api_v1_admin_upload_form_f1_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"},"reporting_period_id":{"type":"string","format":"uuid","title":"Reporting Period Id"}},"type":"object","required":["file","reporting_period_id"],"title":"Body_upload_formf1_api_v1_admin_upload_form_f1_post"},"Body_upload_public_holidays_api_v1_admin_upload_public_holidays_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_public_holidays_api_v1_admin_upload_public_holidays_post"},"Body_upload_rdb_api_v1_admin_upload_rdb_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"},"reporting_period_id":{"type":"string","format":"uuid","title":"Reporting Period Id"}},"type":"object","required":["file","reporting_period_id"],"title":"Body_upload_rdb_api_v1_admin_upload_rdb_post"},"Body_upload_ttf_api_v1_admin_upload_ttf_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"},"reporting_period_id":{"type":"string","format":"uuid","title":"Reporting Period Id"},"programme_code":{"type":"string","title":"Programme Code"}},"type":"object","required":["file","reporting_period_id","programme_code"],"title":"Body_upload_ttf_api_v1_admin_upload_ttf_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status","default":"ok"}},"type":"object","title":"HealthResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}